Make ShardId u16

+++ types/src/lib.rs
@@ -34,4 +34,4 @@ pub mod parcel;
 pub mod transaction;
 
 pub type BlockNumber = u64;
-pub type ShardId = u32;
+pub type ShardId = u16;
